+++ dev/example.py
@@ -1,0 +1,5 @@
+from menus import Engine
+
+if __name__ == '__main__':
+    e = Engine(example=True)
+    e.start()