+++ docs/changelog.md
@@ -1,4 +1,4 @@
-# Changelog
+#Changelog
 
 ## v0.2.0 - Master
 ####* This version is not yet released and is under active development.
@@ -11,6 +11,7 @@
   - A few tests
   - pause method on BaseMenu
     - pause(seconds=5, enter_to_continue=False)
+  - init BaseMenu(commands=commands)
 
 ###Updated
 
@@ -28,6 +29,10 @@
   - Requirement to pass MainMenu option
   - Duplicate code
 
+###Deprecation
+
+  - Initialize BaseMenu with options -> commands
+  - BaseMenu get_correct_action -> get_correct_answer
 
 ##v0.1.0 - 2015/07/25